Computer started glitching, so bad commit
Implemented most of the controller for the lead panel, but
realized not using a table was a mistake(can't recall why I
didn't, but it surely wasn't worth it) and started going back
to the original view design.

Implemented most of the controller for the lead adding page,
missing only some refactoring.

Created a service to handle lead fetching. Only local due to
the scope of the challenge. Should be done